Read and write PSLF 3-winding transformers; consolidate the PSLF writer
PSLF tertiary-winding records were detected and dropped. Map them onto the
Transformer3W type instead: read_transformer now returns a TwoWinding /
ThreeWinding result, building a Transformer3W from the three terminal buses and
the three pairwise impedances (the .epc record carries the primary winding's
ratio and ratings; the secondary and tertiary winding ratios have no column at
these positions, so they default to nominal and the loss is reported). The
writer emits the matching tertiary record after the two-winding ones in the same
transformer section.

Also move the PSLF writer out of the standalone pslf_write.rs and into pslf.rs,
so read and write live in one module like every other format (psse.rs,
egret.rs, powerworld/map.rs). No behavior change from the move.

The Transformer3W impedances keep the Impedance{r, x, base_mva} type: the
per-winding-pair MVA base is declared independently in the source record and is
load-bearing once a non-unit impedance-units convention is read, so it travels
with the value rather than being implied by position.

Tests: a 3-bus .epc with a tertiary record reads to one Transformer3W (not
folded into branches) and round-trips through the writer with its buses,
pairwise impedances, and primary tap intact.
